+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
 All notable changes are documented in this file.
+
+## Unreleased
+
+* Default spell check follows the OS locale instead of the first reverse-sorted dictionary (Turkish)
+
 ## [Version ~1.8.11](https://github.com/asciidocfx/AsciidocFX/releases/tag/v1.8.11) (10.2025)
 
 * Java/FX 25